requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
dparse==0.6.2
idna==3.4
mypy==1.2.0
mypy-extensions==1.0.0
packaging==21.3
paho-mqtt==1.6.1
pyparsing==3.0.9
requests==2.30.0
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
safety==2.3.5
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
yarg==0.1.9
